Moved the green line into a seperate div (read the desc)
I did this to attempt to fix a bug, but it had no effect. I didn't
move it back out of pure laziness. It's 12:33am, so it can stay in
its own element. This is not critical, not just entirely elegant.

So we'll probably regret this commit in the future. Oh well.

Note that the decorator used to be a border-radius on the mainbar.
I moved it initially since it was affecting the height of the bar,
.height() would not take it into account. It turns out .height()
doesn't take padding into account either, so the move, which uses
padding, has literally the same problem.

Damn.