Many changes! Read the desc. [New util, widgets, modified base widget]
Added a base AWidgetSidebarItem, along with a sidebar Object and
ObjectGroup. Also added a new utility to generate unique ids which
are globally valid, along with a helper function to generate those
ids with specific prefixes.

Moved all existing and new classes onto this id system, so updated
the styles and some selectors to reflect this. Added a whole host
of refactorings/optimizations to the issue tracker as a result.

Also updated widget z-indexes