Added manipulatables, basic drag/drop works. See desc
Manipulatable objects share a common global properties class, which
descendants can modify as they wish. When an object is dragged onto
the workspace, it is expected to return the render of a new
manipulatable object.

The properties object will be used to generate a properties form when
the object is selected.